% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasEvaluation.R
\name{anzahlWerteProIndikator}
\alias{anzahlWerteProIndikator}
\title{Anzahl an Werten pro Indikator}
\usage{
anzahlWerteProIndikator(
  effekte = kuras_effekt(),
  indikatoren = kuras_indikator(),
  indikatorwerte = kuras_indikatorwert()
)
}
\value{
data frame mit Spalten \emph{iID} (Indikator-ID), \emph{iEffektID} 
  (Effekt-ID),\emph{iName} (Indikator-Name), \emph{x} (Anzahl Werte fuer den
  Indikator).
}
\description{
Anzahl an Werten pro Indikator
}
