% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasMonitoring.R
\name{generateSampleTimesFromFlaschenbericht}
\alias{generateSampleTimesFromFlaschenbericht}
\title{Generate Sample Times From Flaschenbericht}
\usage{
generateSampleTimesFromFlaschenbericht(
  content,
  text.success = "SUCCESS",
  text.failure = "FAILURE"
)
}
\arguments{
\item{content}{data frame with columns \code{timeFrom}, \code{timeTo},
\code{pnSoll}, \code{pnIst},}

\item{text.success}{text to be put in output column \code{result} in case of
success}

\item{text.failure}{text to be put in output column \code{result} in case of
failure}
}
\description{
Generate equidistant sample times from the informatin given in the
  "Flaschenbericht"
}
