% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasEvaluation.R
\name{kuras_auswertung}
\alias{kuras_auswertung}
\title{kuras_auswertung}
\usage{
kuras_auswertung(
  effekte = kuras_effekt(stringsAsFactors = FALSE),
  indikatoren = kuras_indikator(stringsAsFactors = FALSE),
  indikatorwerte = kuras_indikatorwert(stringsAsFactors = FALSE)
)
}
\description{
kuras_auswertung
}
