% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasMonitoring.R
\name{readKurasSamplerFileByName.R}
\alias{readKurasSamplerFileByName.R}
\title{readKurasSamplerFileByName.R}
\usage{
readKurasSamplerFileByName.R(samplerFile)
}
\description{
read file generated by KURAS_convertAutosamplerFiles_RBFAB.R from 
  "_flaschenbericht" file
}
