% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasMonitoring.R
\name{readKurasSamplerFileByName}
\alias{readKurasSamplerFileByName}
\title{Read KURAS Sampler File by Name}
\usage{
readKurasSamplerFileByName(samplerFile, bottlesToConsider = NA)
}
\arguments{
\item{samplerFile}{path to file written by autosampler}

\item{bottlesToConsider}{arguments passed to
\code{\link{generateSampleTimesFromFlaschenbericht}}}
}
\description{
Read KURAS Sampler File by Name
}
