% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasMonitoring.R
\name{generateSampleTimesFromFlaschenbericht}
\alias{generateSampleTimesFromFlaschenbericht}
\title{generateSampleTimesFromFlaschenbericht}
\usage{
generateSampleTimesFromFlaschenbericht(
  content,
  text.success = "SUCCESS",
  text.failure = "FAILURE"
)
}
\arguments{
\item{content}{data frame with columns \emph{timeFrom}, \emph{timeTo}, \emph{pnSoll},
\emph{pnIst},}
}
\description{
generate equidistant sample times from the informatin given in the
  "Flaschenbericht"
}
