% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasMonitoring.R
\name{readKurasSamplerFileByName.R}
\alias{readKurasSamplerFileByName.R}
\title{Read KURAS Sampler File by Name}
\usage{
readKurasSamplerFileByName.R(samplerFile)
}
\arguments{
\item{samplerFile}{path to file written by autosampler}
}
\description{
Read file generated by KURAS_convertAutosamplerFiles_RBFAB.R from 
  "_flaschenbericht" file
}
