% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kurasMonitoring.R
\name{readKurasSamplerFileByName}
\alias{readKurasSamplerFileByName}
\title{readKurasSamplerFileByName}
\usage{
readKurasSamplerFileByName(samplerFile, bottlesToConsider = NA)
}
\arguments{
\item{bottlesToConsider}{arguments passed to \code{\link{generateSampleTimesFromFlaschenbericht}}}
}
\description{
readKurasSamplerFileByName
}
